Any one give a visha
I am a Bangladeshi at the age of 25, this time I am living very hard. I am the eldest child of the family I am the child of a poor family. I am now a victim of political revenge in Bangladesh, and financially very damaged. The Bangladesh government will not let me live. I am having a hard time with my family. My earnest request to the UK government would be to grant me a shelter and a job visa in this Corona epidemic.